requirements.txt
# FastMCP framework
fastmcp>=2.0.0

# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0

# Web framework dependencies (for HTTP transports)
fastapi>=0.100.0
uvicorn>=0.23.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Logging and monitoring
structlog>=23.0.0
